Remove the Covid service
