--- Guppi.Domain/Interfaces/IWeatherService.cs
@@ -1,12 +1,0 @@
-using System.Threading.Tasks;
-using Guppi.Domain.Entities.Weather;
-
-namespace Guppi.Domain.Interfaces
-{
-    public interface IWeatherService
-    {
-        Task<WeatherForecast> GetWeatherData();
-
-        void Configure();
-    }
-}
