+++ Guppi/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "Guppi.Console": {
       "commandName": "Project",
-      "commandLineArgs": "weather view"
+      "commandLineArgs": "covid view"
     }
   }
 }
